refactoring, rewrite to DDD

+++ src/Application/Service/BirthdayService.go
@@ -1,6 +1,7 @@
 package Service
 
 import (
+	"github.com/umirode/go-rest/src/Domain/Error"
 	"github.com/umirode/go-rest/src/Domain/Model/Entity"
 	"github.com/umirode/go-rest/src/Domain/Repository"
 	"github.com/umirode/go-rest/src/Domain/Service/DTO"
@@ -8,6 +9,12 @@ import (
 
 type BirthdayService struct {
 	birthdayRepository Repository.IBirthdayRepository
+}
+
+func NewBirthdayService(birthdayRepository Repository.IBirthdayRepository) *BirthdayService {
+	return &BirthdayService{
+		birthdayRepository: birthdayRepository,
+	}
 }
 
 func (s *BirthdayService) GetAllForUser(user *Entity.User) ([]*Entity.Birthday, error) {
@@ -19,7 +26,16 @@ func (s *BirthdayService) GetAllForUser(user *Entity.User) ([]*Entity.Birthday, 
 	return users, nil
 }
 
-func (s *BirthdayService) Create(birthdayDTO DTO.BirthdayDTO, user *Entity.User) error {
+func (s *BirthdayService) GetOneById(id uint) (*Entity.Birthday, error) {
+	user, err := s.birthdayRepository.FindOneById(id)
+	if err != nil {
+		return nil, Error.NewNotFoundError()
+	}
+
+	return user, nil
+}
+
+func (s *BirthdayService) Create(birthdayDTO *DTO.BirthdayDTO, user *Entity.User) error {
 	birthday := &Entity.Birthday{
 		Name: birthdayDTO.Name,
 		Date: birthdayDTO.Date,
@@ -35,7 +51,11 @@ func (s *BirthdayService) Create(birthdayDTO DTO.BirthdayDTO, user *Entity.User)
 	return nil
 }
 
-func (s *BirthdayService) Update(birthday *Entity.Birthday, birthdayDTO DTO.BirthdayDTO) error {
+func (s *BirthdayService) Update(birthday *Entity.Birthday, birthdayDTO *DTO.BirthdayDTO, user *Entity.User) error {
+	if birthday.User.ID != user.ID {
+		return Error.NewAccessError()
+	}
+
 	birthday.Name = birthdayDTO.Name
 	birthday.Date = birthdayDTO.Date
 
@@ -48,7 +68,11 @@ func (s *BirthdayService) Update(birthday *Entity.Birthday, birthdayDTO DTO.Birt
 	return nil
 }
 
-func (s *BirthdayService) Delete(birthday *Entity.Birthday) error {
+func (s *BirthdayService) Delete(birthday *Entity.Birthday, user *Entity.User) error {
+	if birthday.User.ID != user.ID {
+		return Error.NewAccessError()
+	}
+
 	err := s.birthdayRepository.Delete(birthday)
 
 	if err != nil {
